+++ demo/src/main/java/com/lin/cms/demo/mapper/BookMapper.java
@@ -8,7 +8,7 @@ public interface BookMapper extends BaseMapper<BookDO> {
 
     BookDO getBookByKeyword(@Param("q") String q);
 
-    BookDO findOneByIdAndDeleteTime(@Param("id") Integer id);
+    BookDO findOneByIdAndDeleteTime(@Param("id") Long id);
 
     BookDO findOneByTitle(@Param("title") String title);
 }
