+++ demo/src/main/java/com/lin/cms/demo/service/AdminService.java
@@ -12,23 +12,23 @@ import java.util.List;
 
 
 public interface AdminService {
-    PageResult getUsers(Integer groupId, Integer count, Integer page);
+    PageResult getUsers(Long groupId, Long count, Long page);
 
-    void changeUserPassword(Integer id, ResetPasswordDTO validator) throws NotFound;
+    void changeUserPassword(Long id, ResetPasswordDTO validator) throws NotFound;
 
-    void deleteUser(Integer id) throws NotFound;
+    void deleteUser(Long id) throws NotFound;
 
-    void updateUserInfo(Integer id, UpdateUserInfoDTO validator) throws NotFound, Parameter;
+    void updateUserInfo(Long id, UpdateUserInfoDTO validator) throws NotFound, Parameter;
 
-    PageResult getGroups(Integer page, Integer count);
+    PageResult getGroups(Long page, Long count);
 
-    GroupWithAuthsBO getGroup(Integer id);
+    GroupWithAuthsBO getGroup(Long id);
 
     void createGroup(NewGroupDTO validator) throws Forbidden;
 
-    void updateGroup(Integer id, UpdateGroupDTO validator) throws NotFound;
+    void updateGroup(Long id, UpdateGroupDTO validator) throws NotFound;
 
-    void deleteGroup(Integer id) throws NotFound, Forbidden;
+    void deleteGroup(Long id) throws NotFound, Forbidden;
 
     void dispatchAuth(DispatchAuthDTO validator) throws NotFound, Forbidden;
