+++ demo/src/test/java/com/lin/cms/demo/api/cms/LogControllerTest.java
@@ -45,7 +45,7 @@ public class LogControllerTest {
     private String method = "GET";
     private String path = "/";
     private Integer statusCode = 200;
-    private Integer userId = 1;
+    private Long userId = 1L;
     private String userName = "pppp";
 
     @Before
