+++ demo/src/test/java/com/lin/cms/demo/api/cms/UserControllerTest.java
@@ -42,7 +42,7 @@ public class UserControllerTest {
 
     private String email = "13129982604@qq.com";
     private String password = "123456";
-    private Integer groupId = 100;
+    private Long groupId = 100L;
     private String nickname = "pedro";
 
     @Before
